fix: wrap sign-up params in suspense
