+++ src/app/(auth)/sign-up/layout.tsx
@@ -1,0 +1,15 @@
+import type { Metadata } from 'next';
+
+export const dynamic = 'force-dynamic';
+
+export const metadata: Metadata = {
+  title: 'Create your heroBooks account',
+};
+
+export default function SignUpLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return children;
+}
